modified:   firmware/application/external/external.cmake 	modified:   firmware/application/external/external.ld 	deleted:    firmware/application/external/test_rx/main.cpp 	deleted:    firmware/application/external/test_rx/ui_test_rx.cpp 	deleted:    firmware/application/external/test_rx/ui_test_rx.hpp 	modified:   firmware/application/ui/ui_spectrum.cpp 	modified:   firmware/application/ui/ui_spectrum.hpp 	modified:   firmware/baseband/CMakeLists.txt 	deleted:    firmware/baseband/proc_test_rx.cpp 	deleted:    firmware/baseband/proc_test_rx.hpp 	modified:   firmware/common/dsp_adv.cpp 	modified:   firmware/common/dsp_adv.hpp 	deleted:    firmware/common/dsp_iq_balancer.cpp 	deleted:    firmware/common/dsp_iq_balancer.hpp 	modified:   firmware/common/spi_image.hpp

+++ firmware/common/dsp_adv.cpp
@@ -181,7 +181,7 @@ float scale_power(float raw, ScaleMode mode, float gain) {
 }
 void generate_lut(ColorScheme scheme) {
     active_scheme = scheme;
-    for (size_t i = 0; i < 256; i++) {
+    for (int i = 0; i < 256; i++) {
         float v = i / 255.0f;
         switch (scheme) {
             case LUT_GRAY:
